Project_Landru/backend/requirements.txt
elevenlabs==0.2.9
pydantic==1.10.13
openai>=1.12.0
flask>=3.0.2
python-dotenv>=1.0.0
flask-cors>=4.0.0